+++ packages/kcms/src/match/routing.ts
@@ -1,0 +1,25 @@
+import { createRoute } from '@hono/zod-openapi';
+import { GetMatchResponseSchema, CommonErrorSchema } from '../match/adaptor/validator/match';
+
+export const GetMatchRoute = createRoute({
+  method: 'get',
+  path: '/match',
+  responses: {
+    200: {
+      content: {
+        'application/json': {
+          schema: GetMatchResponseSchema,
+        },
+      },
+      description: 'Retrieve all matches',
+    },
+    400: {
+      content: {
+        'application/json': {
+          schema: CommonErrorSchema,
+        },
+      },
+      description: 'Common error',
+    },
+  },
+});
